+++ src/@core/utils/validations/validators.js
@@ -1,0 +1,7 @@
+export const validatorIsImageFile = file => {
+  if (!file?.type) {
+    return false
+  }
+  const validTypes = ["image/jpeg", "image/png", "image/webp"];
+  return validTypes.includes(file.type);
+};