+++ src/global-components.js
@@ -1,0 +1,6 @@
+import Vue from 'vue'
+import {ValidationObserver, ValidationProvider} from "vee-validate";
+
+// VeeValidate
+Vue.component('ValidationObserver', ValidationObserver);
+Vue.component('ValidationProvider', ValidationProvider);